Split up shader mod
